[core] enable Async CompletableFuture/CompletionStage interop on native
