--- kyo-core/native/src/main/scala/kyo/internal/AsyncPlatformSpecific.scala
@@ -1,3 +1,0 @@
-package kyo.internal
-
-trait AsyncPlatformSpecific
